pyproject.toml
[tool.poetry]
name = "fin-backend"
version = "0.1.0"
description = ""
authors = ["Ignacio Delmont <ignaciodelmont@gmail.com>"]
readme = "README.md"
packages = [{include = "fin"}]

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = {extras = ["debug-server"], version = "^0.199.2"}
boto3 = "^1.29.6"
python-dateutil = "^2.8.2"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
jinja2 = "^3.1.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.4"
snapshottest = "^0.6.0"
black = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
